Split up different functions for better coverage understanding. added coveragerc to exclude numpy
