+++ .coveragerc:
@@ -1,0 +1,5 @@
+[run]
+omit =
+    */numpy/*
+    */re.py
+
